fix: big refactor of patch.py (#493)
