+++ tests/test_patch.py
@@ -3,7 +3,7 @@ import functools
 from openai import AsyncOpenAI, OpenAI
 
 import instructor
-from instructor.patch import OVERRIDE_DOCS, is_async
+from instructor.utils import is_async
 
 
 def test_patch_completes_successfully():
@@ -71,9 +71,3 @@ def test_is_async_returns_true_if_triple_wrapped_function_is_async():
         pass
 
     assert is_async(triple_wrapped_function) is True
-
-
-def test_override_docs():
-    assert (
-        "response_model" in OVERRIDE_DOCS
-    ), "response_model should be in OVERRIDE_DOCS"
